年版別の計算器を整理
