+++ docs/design-principles.md
@@ -55,8 +55,9 @@
 ## コード構成
 
 - `App.tsx`は現在表示する画面だけを持ち、旧画面をアーカイブ目的で残しません。
-- 路線データと運賃・料金計算はReactコンポーネントから分離し、画面から計算モジュールを利用します。
-- 2022年版の計算コードは参考表示に必要な現役の計算モジュールとして保持し、使用していないUIとは共存させません。
+- `domain/versions/2022.ts`は2022年版の路線・運賃・料金・交換ポイントをまとめた計算器とします。
+- `domain/versions/2026.ts`は2026年版の運賃・料金・交換ポイントをまとめた計算器とします。
+- `domain/quote.ts`で年版ごとの計算器を選択・構成し、Reactコンポーネントには計算式を置きません。
 - 画面内の「区間指定」と「ランキング」はページルーティングではなくタブとして扱い、ホスティングサービス固有のrewriteを必要としない構成にします。
 
 ## 計算根拠
