+++ src/App.test.ts
@@ -1,15 +1,15 @@
 import { describe, expect, it } from "vitest";
 import {
-  legacy2022Engine,
+  calculator2022,
   type Line,
   type SortedSection,
-} from "./legacy2022Engine";
+} from "./domain/versions/2022";
 import {
-  createQuote,
   intervalWithin,
   rangeAfterEndChange,
   rangeAfterStartChange,
 } from "./App";
+import { createQuote } from "./domain/quote";
 
 const section = (line: Line, departure: string, arrival: string): SortedSection => ({
   departure: line.find((station) => station.name === departure)!,
@@ -18,7 +18,7 @@ const section = (line: Line, departure: string, arrival: string): SortedSection 
 });
 
 const route = (group: string, index = 0) =>
-  legacy2022Engine.lineGroups.get(group)!.lines[index]!;
+  calculator2022.lineGroups.get(group)!.lines[index]!;
 
 describe("2026年版の通し特急料金", () => {
   it("宇都宮―那須塩原のポイントと比較額を維持する", () => {
@@ -28,7 +28,7 @@ describe("2026年版の通し特急料金", () => {
       campaign: "regular",
       line,
       section: section(line, "宇都宮", "那須塩原"),
-      season: legacy2022Engine.average,
+      season: calculator2022.average,
     });
 
     expect(quote.points).toBe(2_000);
@@ -43,7 +43,7 @@ describe("2026年版の通し特急料金", () => {
       campaign: "regular",
       line,
       section: section(line, "東京", "新庄"),
-      season: legacy2022Engine.busy,
+      season: calculator2022.busy,
     });
 
     expect(quote.basicFare).toBe(7_480);
@@ -60,7 +60,7 @@ describe("2026年版の通し特急料金", () => {
       line,
       section: trip,
       green: { start: trip.departure.index, end: trip.arrival.index },
-      season: legacy2022Engine.busy,
+      season: calculator2022.busy,
     });
 
     expect(quote.expressFare).toBe(5_520);
@@ -75,7 +75,7 @@ describe("2026年版の通し特急料金", () => {
       campaign: "regular",
       line,
       section: section(line, "福島", "新庄"),
-      season: legacy2022Engine.busy,
+      season: calculator2022.busy,
     });
 
     expect(quote.expressFare).toBe(2_310);
@@ -88,7 +88,7 @@ describe("2026年版の通し特急料金", () => {
       campaign: "regular",
       line,
       section: section(line, "東京", "ガーラ湯沢"),
-      season: legacy2022Engine.average,
+      season: calculator2022.average,
     });
 
     expect(quote.expressFare).toBe(3_480);
@@ -101,7 +101,7 @@ describe("2026年版の通し特急料金", () => {
       campaign: "regular",
       line,
       section: section(line, "水沢江刺", "秋田"),
-      season: legacy2022Engine.average,
+      season: calculator2022.average,
     });
 
     expect(quote.basicFare).toBe(3_850);
@@ -115,7 +115,7 @@ describe("2026年版の通し特急料金", () => {
       line,
       section: section(line, "東京", "仙台"),
       green: { start: 0, end: line.length },
-      season: legacy2022Engine.average,
+      season: calculator2022.average,
     });
 
     expect(quote.points).toBeUndefined();
@@ -147,7 +147,7 @@ describe("信州プレDCの画面用計算", () => {
       campaign: "shinshuPreDc",
       line,
       section: section(line, "東京", "長野"),
-      season: legacy2022Engine.average,
+      season: calculator2022.average,
     });
     const greenQuote = createQuote({
       version: "2026-03-14",
@@ -158,7 +158,7 @@ describe("信州プレDCの画面用計算", () => {
         start: line.find((station) => station.name === "東京")!.index,
         end: line.find((station) => station.name === "長野")!.index,
       },
-      season: legacy2022Engine.average,
+      season: calculator2022.average,
     });
 
     expect(quote.points).toBe(4_500);
